backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
passlib[bcrypt]
pydantic

pip install -r requirements.txt